(Minor) Clean up some code in Player and Projectile scripts
NOTE: Much more could be cleaned up, and Player and Projectile decoupled, using even more events on the queue. However, it is a risk to enter into such changes both technically and regarding readability, therefore the old code will be kept as is (I hereby give a guarantee that it is coupled but works reliably so far, for the parts that I wanted to replace).